split user tests into their own file
